+++ CalCalculatorEntities/Models/TempFood.cs
@@ -9,7 +9,7 @@ namespace CalCalculatorEntities.Models
     public class TempFood
     {
         public string FoodName { get; set; }
-        public int Calories { get; set; }
+        public double Calories { get; set; }
         public double Grams { get; set; }
         public int FoodID { get; set; }
